Refactor: view 클래스 도메인에 따라 분리
