+++ src/index.js
@@ -1,7 +1,19 @@
 import RacingGameController from "./class/RacingGameController.js";
+import RacingGameModel from "./class/RacingGameModel.js";
+import RacingGameView from "./Views/RacingGameView.js";
+import RacingInputView from "./Views/RacingInputView.js";
+import RacingResultView from "./Views/RacingResultView.js";
 
 document.addEventListener("DOMContentLoaded", racingGame);
 
 function racingGame() {
-  new RacingGameController();
+  const racingGameModel = new RacingGameModel();
+
+  const views = {
+    racingInputView: new RacingInputView(),
+    racingGameView: new RacingGameView(),
+    racingResultView: new RacingResultView(),
+  }
+
+  new RacingGameController(racingGameModel, views);
 }
